Add note on value classes vs @newtype
I think that this piece of information is very crucial since value classes are native for scala. I know that there is this issue https://github.com/estatico/scala-newtype/issues/42 but I would expect to see that information in the top of the readme.

I mostly copy-pasted what was in the issue and made it more discover-able.